+++ Runtime/Framework/StateMachine/StateMachine.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 
 namespace DT.UniStart {
+  [Serializable]
   public class StateMachine<T> : Watch<T>, IStateMachine<T> where T : Enum {
     static readonly T[] values = Enum.GetValues(typeof(T)) as T[];
     readonly Dictionary<T, UniEvent<T, T>> onEnter = new();